# arcsi/static/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Arcsi API doc",
    "version": "1.0.1",
    "title": "Arcsi",
    "contact": {
      "email": "it@lahmacun.hu"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "GuestApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authentication-Token"
      },
      "AdminApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authentication-Token"
      }
    }
  },
  "servers": [
    {
      "url": "/arcsi/"
    }
  ],
...
  "paths": {
...
    "/archon/show/all": {
      "get": {
        "tags": [
          "Show Requests for Archon"
        ],
        "summary": "Return all Shows for Archon's Shows Page",
        "security": [
          {
            "AdminApiKeyAuth": []
          }
        ],
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "show-component.json#/components/schemas/ArchonShowsSchema"
                }
              }
            }
          }
        }
      }
    },
    "/archon/show/{id}": {
      "get": {
        "tags": [
          "Show Requests for Archon"
        ],
        "summary": "Return chosen Show for Archon's Show Subpage",
        "security": [
          {
            "GuestApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Show.id",
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "show-component.json#/components/schemas/ShowSchema"
                }
              }
            }
          },
          "404": {
            "description": "Show not found"
          }
        }
      }
    },
    "/archon/show/add": {
      "post": {
        "tags": [
          "Show Requests for Archon"
        ],
        "summary": "Add Show to Archon (internal use)",
        "security": [
          {
            "AdminApiKeyAuth": []
          }
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "show-component.json#/components/schemas/ShowRequestSchema"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "show-component.json#/components/schemas/ShowSchema"
                }
              }
            }
          },
          "500": {
            "description": "Invalid data sent to add show"
          },
          "503": {
            "description": "Only accepts multipart/form-data for now, sorry"
          }
        }
      }
    },
    "/archon/show/{id}/edit": {
      "post": {
        "tags": [
          "Show Requests for Archon"
        ],
        "summary": "Edit Show in Archon (internal use)",
        "security": [
          {
            "AdminApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Show.id",
            "type": "string"
          }
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "show-component.json#/components/schemas/ShowRequestSchema"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "show-component.json#/components/schemas/ShowSchema"
                }
              }
            }
          },
          "404": {
            "description": "Show not found"
          },
          "500": {
            "description": "Invalid data sent to edit show"
          }
        }
      }
    },
    "/show/list": {
      "get": {
        "tags": [
          "Show Requests for external usage"
        ],
        "summary": "Return all Shows for frontend/app Shows Page",
        "security": [
          {
            "GuestApiKeyAuth": []
          }
        ],
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "show-component.json#/components/schemas/ShowsArchiveSchema"
                }
              }
            }
          }
        }
      }
    },
    "/show/all_without_items": {
      "get": {
        "tags": [
          "Show Requests for external usage"
        ],
        "summary": "Return all Shows for frontend/app Shows Page",
        "security": [
          {
            "GuestApiKeyAuth": []
          }
        ],
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "show-component.json#/components/schemas/ShowsScheduleExcludedSchema"
                }
              }
            }
          }
        }
      }
    },
    "/show/all_details": {
      "get": {
        "tags": [
          "Show Requests for external usage"
        ],
        "summary": "Return all Shows for frontend/app Shows Page",
        "security": [
          {
            "GuestApiKeyAuth": []
          }
        ],
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "show-component.json#/components/schemas/ShowsDetailsSchema"
                }
              }
            }
          }
        }
      }
    },
    "/show/all_schedule?week={week}": {
      "get": {
        "tags": [
          "Show Requests for external usage"
        ],
        "summary": "Return all Shows for frontend/app Schedule Page",
        "security": [
          {
            "GuestApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "week",
            "required": false,
            "description": "week, default: 1",
            "type": "number"
          }
        ],
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "show-component.json#/components/schemas/ShowsScheduleSchema"
                }
              }
            }
          }
        }
      }
    },
    "/show/search?size={size}&page={page}&param={param}": {
      "get": {
        "tags": [
          "Show Requests for external usage"
        ],
        "summary": "Return all searched Shows (their name or description contains the parameter)",
        "security": [
          {
            "GuestApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "size",
            "required": false,
            "description": "page_size, default: 12",
            "type": "integer"
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "description": "page_number, default: 1",
            "type": "integer"
          },
          {
            "in": "query",
            "name": "param",
            "required": false,
            "description": "search_parameter, default: 'lahmacun'",
            "type": "string"
          }
        ],
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "show-component.json#/components/schemas/ShowsScheduleExcludedSchema"
                }
              }
            }
          }
        }
      }
    },
    "/show/search_by_tag/{clean_tag}": {
      "get": {
        "tags": [
          "Show Requests for external usage"
        ],
        "summary": "Return all Shows with given Tag",
        "security": [
          {
            "GuestApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clean_tag",
            "required": true,
            "description": "Tag.clean_name",
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "show-component.json#/components/schemas/ShowsDetailsSchema"
                }
              }
            }
          }
        }
      }
    },
    "/show/schedule?week={week}": {
      "get": {
        "tags": [
          "Show Requests for external usage"
        ],
        "summary": "Return all active Shows for frontend/app Home and Schedule Page",
        "security": [
          {
            "GuestApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "week",
            "required": false,
            "description": "week, default: 1",
            "type": "number"
          }
        ],
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "show-component.json#/components/schemas/ShowsScheduleExcludedSchema"
                }
              }
            }
          }
        }
      }
    },
    "/show/schedule_by?week={week}&day={day}": {
      "get": {
        "tags": [
          "Show Requests for external usage"
        ],
        "summary": "Return all active Shows airing on the chosen day for frontend/app Home and Schedule Page",
        "security": [
          {
            "GuestApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "week",
            "required": false,
            "description": "week, default: 1",
            "type": "number"
          },
          {
            "in": "query",
            "name": "day",
            "required": false,
            "description": "day, default: 1",
            "type": "number"
          }
        ],
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "show-component.json#/components/schemas/ShowsScheduleSchema"
                }
              }
            }
          }
        }
      }
    },
    "/show/{show_slug}/page?filter={filter}": {
      "get": {
        "tags": [
          "Show Requests for external usage"
        ],
        "summary": "Return the chosen Show for frontend/app Show Subpage",
        "security": [
          {
            "GuestApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "show_slug",
            "required": true,
            "description": "Show.archive_lahmastore_base_url",
            "type": "string"
          },
          {
            "in": "query",
            "name": "filter",
            "required": false,
            "description": "can be archived (sorting out upcoming episodes) and/or latest (return only the latest)",
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "show-component.json#/components/schemas/ShowArchivePageSchema"
                }
              }
            }
          },
          "404": {
            "description": "Show not found"
          }
        }
      }
    },
...
    "/show/{show_slug}/item/{episode_slug}": {
      "get": {
        "tags": [
          "Item Requests for external usage"
        ],
        "summary": "Return Item for frontend/app Episode Subpage",
        "security": [
          {
            "GuestApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "show_slug",
            "required": true,
            "description": "Show.archive_lahmastore_base_url",
            "type": "string"
          },
          {
            "in": "path",
            "name": "episode_slug",
            "required": true,
            "description": "normalize(Item.name) --> name_slug",
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "item-component.json#/components/schemas/ItemArchiveSchema"
                }
              }
            }
          },
          "404": {
            "description": "Episode not found"
          }
        }
      }
    },
...
  }
}